Add parsing of top-level objects.
By parsing the top-level objects of a Kotlin file, we can list a comprehensive
list of Kotlin identifiers (i.e. fully qualified names) for the source files
that are part of a given Bazel target. If a Bazel package has multiple targets
defined, we can extract the top-level identifiers of each target by analyzing
the source file members of each target. These fine-grained identifiers can be
used as the basis for import resolution instead of using the package name.